+++ .gitignore
@@ -1,0 +1,6 @@
+*.xml
+*.bin
+*.mp4
+*.jpg
+*.png
+*.weights
