--- class01/homework/WooChangMin/hw1_Day07_OpenCV/opencv/quiz
@@ -1,37 +1,0 @@
-
-2-1.
-2-2.
-2-3.
-2-4.
-
-
-
-
-3-1. 색공간 표현방법 이해->완료
-3-2. HSV가 효과적으로 사용될 때->어두울때
-3-3. HSV변환 이미지를 다시 RGB로 변환해서 출력->rgb=cv2.cvtColor(hsv,cv2.COLOR_HSV2RGB)
-3-4. RGB를 다시 흑백으로 변환해서 출력->gray=cv2.cvtColor(color,cv2.COLOR_RGB2GRAY)
-
-
-3-5. input image size 확인 -> print(img.shape)
-3-6. 본인 사용한 이미지 얼굴 영역만 crop 해서 display->cropped = img[130:355,50:395]
-3-7. 원본 이미지 1.5배 확대해서 파일로 저장->half_doubled=cv2.resize(img,(678,1016))   cv2.imshow("half_doubled Image",half_doubled)   cv2.imwrite("half_doubled.png", half_doubled)
-3-8. opencv rotate api사용해서 우측으로 90도 만큼 회전된 이미지 출력
-
-
-4-1. 동영상이 빠르게 재생됨-> 해결방법: key = cv2.waitKey(33)#33ms만큼 대기 1초가 1000ms->
-4-2. 동영상이 끝까지 재생되고 다시 처음부터 재생되도록 수정: ->코드 참고
-4-3. 동영상 크기 반으로 resize해서 출력:->코드 참고
-4-4. 동영상 재생 중 c입력 받으면 frame img 파일로 저장하게 수정 001.jpg 002.jpg 로 overwrite되지 않게:
-
-
-5-1. 해상도 변경 cap.set(cv2.CAP_PROP_FRAME_WIDTH, 640) cap.set(cv2.CAP_PROP_FRAME_HEIGHT, 480) cap.set(cv2.CAP_PROP_FPS, 30)
-5-2. 웹캠 input 동영상 파일로 저장->코드 참고
-
-6-1. text문구 /font/ 색상/ 크기 / 굵기 / 출력위치 등 모든 값 변경->코드 참고
-6-2. 동그라미 그리는 함수->코드 참고
-6-3. 마우스 왼쪽 버튼을 click하면 해당 위치에 동그라미 reference cv2.EVENT_LBUTTONDOWN->코드 참고
-
-6-4. trackbar를 control해서 text 굵기가 변하는 것을 확인->코드 참고
-6-5. trackbar를 추가해서 font size 변경 및 적용->코드 참고
-6-6. R G B trackbar를 각각 추가 글자의 font color를 변경->코드 참고
