+++ class01/homework/WooChangMin/hw2_Day07_OpenCV/opencv/Readme.md
@@ -1,0 +1,8 @@
+
+# Project OpenCV_woochangmin
+
+## OpenCV HomeWork
+
+```
+.py파일들과 quiz를 모아놓은 quiz.txt로 구성되어 있습니다.
+```
