+++ class01/homework/WooChangMin/hw2_Day07_OpenCV/opencv/webcam.py
@@ -1,0 +1,39 @@
+import numpy as np
+import cv2
+
+
+# Read from the recorded video file
+cap = cv2.VideoCapture(0)
+cap.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
+cap.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
+cap.set(cv2.CAP_PROP_FPS, 30)
+
+
+
+# Get the original video's width, height, and frames per second (fps)
+width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
+height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
+
+fps = cap.get(cv2.CAP_PROP_FPS)
+
+
+
+
+print(width,height,fps)
+# 동영상 파일이 성공적으로 열렸으면 while문 반복
+while(cap.isOpened()):
+    # 한 프레임을 읽어옴
+    ret, frame = cap.read()
+    
+    if ret is False:
+        print("Can't receive frame (stream end?). Exiting ...")
+        break
+    # Display
+    cv2.imshow("Camera", frame)
+    
+
+    # i ms 동안 대기하며 키 입력을 받고 'q' 입력 시 종료
+    key = cv2.waitKey(1)#33ms만큼 대기 1초가 1000ms->
+    if key & 0xFF == ord('q'):
+        break
+
