--- class01/homework/woochangmin/README.md
@@ -1,1 +1,0 @@
-woochangmin 1
